Update Vancouver.XSL
Pages are an important field when citing journal articles, also the "p. " it is not used anymore

The actual format is:
Huded CP, Tuzcu EM, Krishnaswamy A, et al. Association Between Transcatheter Aortic Valve Replacement and Early Postprocedural Stroke. JAMA. 2019;321(23):2306–2315

Where the: 2306-2315 are the pages of the journal article.

I believe this would be solved deleting the "p." in line 227